Redesign Home Page Continuation
