+++ app/globals.css
@@ -70,6 +70,14 @@
   .content-visibility-auto {
     content-visibility: auto;
   }
+
+  p.hyphenate {
+    overflow-wrap: break-word;
+    hyphens: auto;
+
+    /* Prefix for Safari */
+    -webkit-hyphens: auto;
+  }
 }
 
 @layer base {
