+++ components/pages/IconButton.tsx
@@ -27,7 +27,7 @@ export default function IconButton({
         href={link}
         className={`flex items-center py-1 rounded-lg text-lg cursor-pointer border-none transition-colors duration-200 font-oswald font-bold space-x-2 px-2 ${
           isActive
-            ? "text-transparent bg-gradient-to-r from-blue-500 to-purple-500 text-white"
+            ? "text-transparent bg-gradient-to-r from-blue-500/80 to-purple-500/80 text-white"
             : "text-white"
         }`}
         onClick={() => callback()}
